Switch to use edown 0.5
We don't want to just point to HEAD, since who knows what we will get.

Torben had previously run into problems already with HEAD, and so he
created a 1.5-basho1 version that used edown 0.7. However, this
conflicts with parse_trans and sext which are both currently specifying
edown 0.5 in develop-2.2. Since 0.5 also seems to work fine for setup,
I'm going to create a 1.5-basho2 tag on this commit which we can use in
Riak going forward.